requirements.txt
pygame
python-dotenv
google-generativeai
numpy
